Add Gravatar for Contact photo if avatar does not exist (#66)
* Add Gravatar for Contact photo if avatar does not exist

Uses Gravatar https://en.gravatar.com/ for profile photo if an existing
avatar does not already exist and the user has an email address. One
thing is that it does not check Gravatar if it exists there, so if email
exists and no Gravatar, the default becomes the Gravatar default and not
the colored initials.

It is possible to check if the Gravatar exists by passing ?d=404 and
querying the headers to see if it returns a 404, but would be an extra
call on each page load, which isn't horrible since only one image per
page.

* Check if Gravatar exists and default to initials if not

Adds a check to see if Gravatar exists by fetching the headers and
seeing if it returns a 404. If Gravatar does not exist defaults to
initials avatar.

Also, image size being displayed is 87 pixels, so we fetch the Gravatar
at 2x so it displays nicely for high dpi displays.