Fix alphabetic country list (#130)
I've added a fix for the alphabetic country list, and externalised the country dropdown to a component. This component is powered by a View composer which means that no data tie up is needed in the main controller..Just plonk the include file in the right place and tell it what it should bind to and away you go. There may be other parts of the project that could benefit from using this component now it is in place.